ask to install arg_scanner gem on execution; pass src roots to gem
The way gem is installed is almost the same as it is for debugger gems.
The difference is that we don't bundle arg_scanner gem with the plugin.

Modules content roots are passed to the gem to distinguish them from
"fixed" gems code. See b52eff2e76d5408353099064b03938bd584ceaf8.